# docs/swagger.yaml
basePath: /api
definitions:
...
  apierror.ApiError:
    properties:
      code:
        type: integer
      error: {}
    type: object
...
  db.Grade:
    properties:
      assignment_id:
        type: integer
      feedback:
        $ref: '#/definitions/pgtype.Text'
      grade:
        type: integer
      graded_at:
        $ref: '#/definitions/pgtype.Timestamp'
      graded_by:
        $ref: '#/definitions/pgtype.Int4'
      id:
        type: integer
      task_id:
        type: integer
      user_id:
        type: integer
    type: object
...
  models.CreateGradeDTO:
    properties:
      assignment_id:
        type: integer
      feedback:
        maxLength: 1000
        type: string
      grade:
        maximum: 5
        minimum: 1
        type: integer
      task_id:
        type: integer
      user_id:
        type: integer
    required:
    - assignment_id
    - grade
    - task_id
    - user_id
    type: object
...
  models.GradeResponse:
    properties:
      assignment_id:
        type: integer
      assignment_title:
        type: string
      feedback:
        $ref: '#/definitions/pgtype.Text'
      grade:
        type: integer
      graded_at:
        $ref: '#/definitions/pgtype.Timestamp'
      graded_by:
        $ref: '#/definitions/pgtype.Int4'
      graded_by_username:
        type: string
      id:
        type: integer
      student_username:
        type: string
      task_id:
        type: integer
      task_title:
        type: string
      user_id:
        type: integer
    type: object
  models.GradeStats:
    properties:
      average_grade:
        type: number
      grade_distribution:
        additionalProperties:
          type: integer
        type: object
      highest_grade:
        type: integer
      lowest_grade:
        type: integer
    type: object
...
  pgtype.InfinityModifier:
    enum:
    - 1
    - 0
    - -1
    type: integer
    x-enum-varnames:
    - Infinity
    - Finite
    - NegativeInfinity
  pgtype.Int4:
    properties:
      int32:
        type: integer
      valid:
        type: boolean
    type: object
  pgtype.Text:
    properties:
      string:
        type: string
      valid:
        type: boolean
    type: object
  pgtype.Timestamp:
    properties:
      infinityModifier:
        $ref: '#/definitions/pgtype.InfinityModifier'
      time:
        description: Time zone will be ignored when encoding to PostgreSQL.
        type: string
      valid:
        type: boolean
    type: object
...
host: localhost:4000
info:
  contact: {}
  title: Rankode
  version: "1.0"
paths:
...
  /grades:
    get:
      consumes:
      - application/json
      description: Retrieves grades based on filters (teacher sees all, students see
        own)
      parameters:
      - description: Filter by assignment ID
        in: query
        name: assignment_id
        type: integer
      - description: Filter by course ID
        in: query
        name: course_id
        type: integer
      - description: Filter by student ID (teacher only)
        in: query
        name: student_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of grades
          schema:
            items:
              $ref: '#/definitions/models.GradeResponse'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/apierror.ApiError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apierror.ApiError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierror.ApiError'
      security:
      - ApiKeyAuth: []
      summary: List grades
      tags:
      - Grades
    post:
      consumes:
      - application/json
      description: Creates or updates a grade for a student's task in an assignment
        (teacher only)
      parameters:
      - description: Grade payload
        in: body
        name: grade
        required: true
        schema:
          $ref: '#/definitions/models.CreateGradeDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Grade created/updated successfully
          schema:
            $ref: '#/definitions/db.Grade'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/apierror.ApiError'
        "403":
          description: Forbidden (not teacher of course)
          schema:
            $ref: '#/definitions/apierror.ApiError'
        "404":
          description: Assignment, task, or student not found
          schema:
            $ref: '#/definitions/apierror.ApiError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierror.ApiError'
      security:
      - ApiKeyAuth: []
      summary: Create or update a grade
      tags:
      - Grades
  /grades/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a grade by ID (teacher only)
      parameters:
      - description: Grade ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Grade deleted successfully
        "403":
          description: Forbidden (not teacher)
          schema:
            $ref: '#/definitions/apierror.ApiError'
        "404":
          description: Grade not found
          schema:
            $ref: '#/definitions/apierror.ApiError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierror.ApiError'
      security:
      - ApiKeyAuth: []
      summary: Delete a grade
      tags:
      - Grades
  /grades/stats:
    get:
      consumes:
      - application/json
      description: Retrieves statistics for an assignment (teacher only)
      parameters:
      - description: Assignment ID
        in: query
        name: assignment_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Grade statistics
          schema:
            $ref: '#/definitions/models.GradeStats'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/apierror.ApiError'
        "403":
          description: Forbidden (not teacher of course)
          schema:
            $ref: '#/definitions/apierror.ApiError'
        "404":
          description: Assignment not found
          schema:
            $ref: '#/definitions/apierror.ApiError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierror.ApiError'
      security:
      - ApiKeyAuth: []
      summary: Get grade statistics
      tags:
      - Grades
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"